+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Percursus/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Percursus/releases)
 
+- Added: Race times for the area 'K'aresh'
 - Changed: Output of debug messages
 - Updated: Race times for the event 'Eastern Kingdom Cup'